requirements.txt
pip
setuptools
wheel
pytz
pyOpenSSL
ndg-httpsclient
pyasn1
redis~=3.2
apache-airflow[statsd,password,crypto,celery,postgres,ssh,docker]==2.2.4 --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.2.4/constraints-no-providers-2.2.4.txt"
virtualenv
statsd
boto3
psycopg2-binary
sqlalchemy>=1.3.18, <1.4
openapi-schema-validator<0.2.0
markupsafe<2.1.0
datetime
authlib
openpyxl
google.cloud
google.oauth
google-auth-oauthlib
google-cloud-bigquery
pandas
numpy
pandas_gbq
pyarrow
gspread
Office365-REST-Python-Client
office365
requests
simple_salesforce
facebook_business